# docs/api/swagger.json
{
	"swagger" : "2.0",
	"info" : {
		"description" : " You can find out more about Microsoft DUCK at [https://github.com/Microsoft/DUCK](https://github.com/Microsoft/DUCK).",
		"version" : "1.0.0",
		"title" : "Microsoft DUCK",
		"termsOfService" : "http://swagger.io/terms/",
		"contact" : {
			"email" : "apiteam@swagger.io"
		},
		"license" : {
			"name" : "Apache 2.0",
			"url" : "http://www.apache.org/licenses/LICENSE-2.0.html"
		}
	},
	"host" : "localhost:8000",
	"basePath" : "/v1",
...
	"schemes" : ["http"],
	"paths" : {
...
		"/users/{userID}" : {
			"delete" : {
				"tags" : ["user"],
				"summary" : "Delete a user",
				"operationId" : "deleteUser",
				"produces" : ["application/json"],
				"parameters" : [{
						"name" : "userID",
						"in" : "path",
						"description" : "ID of user to delete",
						"required" : true,
						"type" : "string"
					}
				],
				"responses" : {
					"200" : {
						"description" : "Successful operation",
						"schema" : {
							"type" : "array",
							"items" : {
								"$ref" : "#/definitions/ApiResponse"
							}
						}
					},
					"404" : {
						"description" : "Invalid User ID",
						"schema" : {
							"$ref" : "#/definitions/ApiResponse"

						}
					}
				},
				"security" : [{
						"jwt_key" : []
					}
				]
			}
		},
...
		"/documents/{documentID}" : {
			"delete" : {
				"tags" : ["document"],
				"summary" : "Delete a document",
				"operationId" : "deleteDocument",
				"produces" : ["application/json"],
				"parameters" : [{
						"name" : "documentID",
						"in" : "path",
						"description" : "ID of document to delete",
						"required" : true,
						"type" : "string"
					}
				],
				"responses" : {
					"200" : {
						"description" : "Successful operation",
						"schema" : {
							"type" : "array",
							"items" : {
								"$ref" : "#/definitions/ApiResponse"
							}
						}
					},
					"404" : {
						"description" : "Invalid document ID",
						"schema" : {
							"$ref" : "#/definitions/ApiResponse"

						}
					}
				},
				"security" : [{
						"jwt_key" : []
					}
				]
			},
			"get" : {
				"tags" : ["document"],
				"summary" : "Delete a document",
				"operationId" : "getDocument",
				"produces" : ["application/json"],
				"parameters" : [{
						"name" : "documentID",
						"in" : "path",
						"description" : "ID of document to return",
						"required" : true,
						"type" : "string"
					}
				],
				"responses" : {
					"200" : {
						"description" : "Successful operation",
						"schema" : {
							"type" : "array",
							"items" : {
								"$ref" : "#/definitions/ApiResponse"
							}
						}
					},
					"404" : {
						"description" : "Invalid User ID",
						"schema" : {
							"$ref" : "#/definitions/ApiResponse"

						}
					}
				},
				"security" : [{
						"jwt_key" : []
					}
				]
			}
		},
...
	},
	"securityDefinitions" : {
		"jwt_key" : {
			"type" : "apiKey",
			"description" : "A JSON Web Token",
			"name" : "jwt_key",
			"in" : "header"
		}
	},
	"definitions" : {
		"ApiResponse" : {
			"type" : "object",
			"required" : ["ok"],
			"properties" : {
				"ok" : {
					"type" : "boolean"
				},
				"reason" : {
					"type" : "string",
					"description" : "Reason is given when ok is false"
				}
			}
		},
...
	}
}